List instead of model

+++ Rates.API/Rates.API/RatesModel.cs
@@ -1,9 +1,9 @@
 ﻿namespace Rates.API
 {
     public class RatesModel
-    {       
-        public double USD { get; set; }
-        public double RUB { get; set; }
-        public double JPY { get; set; }
+    {
+        public decimal USD { get; set; }
+        public decimal RUB { get; set; }
+        public decimal JPY { get; set; }
     }
 }
